RoyaleRL/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
asttokens==3.0.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
easyocr==1.7.2
executing==2.2.0
filelock==3.19.1
fonttools==4.59.1
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
imageio==2.37.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.1
parso==0.8.4
pillow==11.3.0
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
PyAutoGUI==0.9.54
pyclipper==1.3.0.post6
pycocotools==2.0.10
pydantic==2.11.7
pydantic_core==2.33.2
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
scikit-image==0.25.2
scipy==1.16.1
seaborn==0.13.2
setuptools==80.9.0
shapely==2.1.1
simsimd==6.5.1
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
stringzilla==3.12.6
sympy==1.14.0
tifffile==2025.6.11
torch==2.3.1+cu121
torchvision==0.18.1+cu121
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.182
ultralytics-thop==2.0.16
urllib3==2.5.0
wcwidth==0.2.13
wheel==0.45.1